Added autoRouter to avoid routing issues. Now all modules exposing an express.Router() will be mapped automatically according to their location relative to the routes folder
For instnace, the `register.js` route, which is located in `/routes/user` folder, is automatically mapped to `/user/register` route.

With this mechanism .js files that handle different resources need to only be placed in the appropriate directories and they will be picked up automatically, moving them will also update their routes.